requirements.txt
Django==3.1.4
python-dotenv==0.15.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
